backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
email-validator==1.1.3
python-dotenv==0.19.0
sqlalchemy==1.4.23
bcrypt==4.0.1
google-auth-oauthlib==1.0.0
google-api-python-client==2.86.0
